requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.2
django-cors-middleware==1.5.0
django-extensions==3.2.1
djangorestframework==3.14.0
PyJWT==2.6.0
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
